refactor: 문자열 콤마 분리 메소드 분리
