+++ src/main/java/racingcar/datatransfer/CarNamesRequest.java
@@ -1,19 +1,14 @@
-package racingcar.dto;
+package racingcar.datatransfer;
 
-import java.util.Arrays;
 import java.util.List;
-import java.util.stream.Collectors;
+
+import static racingcar.utils.SplitStringByComma.splitByComma;
 
 public class CarNamesRequest {
     private final List<String> carNames;
 
     public CarNamesRequest(String carNames) {
         this.carNames = splitByComma(carNames);
-    }
-
-    private List<String> splitByComma(String carNames) {
-        String[] splitCarNames = carNames.split("\\s*,\\s*");
-        return Arrays.asList(splitCarNames);
     }
 
     public List<String> getCarNames() {
